refactor(convention): global 디렉터리 생성
