+++ src/global/constants/exception.constant.ts
