+++ src/global/types/SVGAttribute.ts
