+++ src/page/applications/constants/applications.constant.ts
