+++ src/store/categories.store.ts
@@ -1,4 +1,4 @@
-import forum from "@/utils/constants/forum.constant";
+import forum from "@/page/forum/constants/forum.constant";
 import { atom } from "recoil";
 
 export const categoriesStore = atom<string>({
